validacion de cuit
incompleta

+++ SistemaDeStock/Cliente.h
@@ -7,18 +7,18 @@ class Cliente{
 private:
     int _ID;
     char _NombreEmpresa[50];
-    double _Cuit;
+    int _Cuit;
     char _Correo[50];
     bool _Estado;
 public:
     void setID(int id);
     void setNombreEmpresa(string NombreEmpresa);
-    void setCuit(double Cuit);
+    void setCuit(int Cuit);
     void setCorreo(string Correo);
     void setEstado(bool Estado);
     int getID();
     string getNombreEmpresa();
-    double getCuit();
+    int getCuit();
     string getCorreo();
     bool getEstado();
     void Cargar();
